+++ banking-api-tewzka/src/main/java/com/ws/tla/model/dto/CardDto.java
@@ -14,5 +14,5 @@ public class CardDto {
     private CardType cardType;
     private String cardNumber;
     private Long cvvNumber;
-    private Long availableCardLimit;
+    private Integer availableCardLimit;
 }
